# docs/swagger.yaml
definitions:
  handlers.Bid:
    properties:
      comments:
        type: string
      contractor_id:
        type: string
      delivery_time:
        type: integer
      id:
        type: string
      price:
        type: number
      status:
        type: string
      tender_id:
        type: string
    type: object
  handlers.CreateBidRequest:
    properties:
      comments:
        type: string
      delivery_time:
        type: integer
      price:
        type: number
    type: object
...
  handlers.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend
  title: hackathon
  version: "1.0"
paths:
...
  /api/client/tenders/{tender_id}/award/{bid_id}:
    post:
      consumes:
      - application/json
      description: This endpoint allows a client to award a specific bid for a specified
        tender.
      parameters:
      - description: Tender ID
        in: path
        name: tender_id
        required: true
        type: string
      - description: Bid ID
        in: path
        name: bid_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully awarded bid
          schema:
            $ref: '#/definitions/handlers.Bid'
        "400":
          description: Invalid tender ID or bid ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Award a bid
      tags:
      - bids
  /api/client/tenders/{tender_id}/bids:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves a list of bids made by a specific client
        for a specified tender.
      parameters:
      - description: Tender ID
        in: path
        name: tender_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of bids
          schema:
            items:
              $ref: '#/definitions/handlers.Bid'
            type: array
        "400":
          description: Invalid tender ID or client ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get bids by client ID
      tags:
      - bids
...
  /api/contractor/bids:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves a list of bids made by a specific contractor.
      parameters:
      - description: Contractor ID
        in: path
        name: contractor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of bids
          schema:
            items:
              $ref: '#/definitions/handlers.Bid'
            type: array
        "400":
          description: Invalid contractor ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get bids by contractor ID
      tags:
      - bids
  /api/contractor/bids/{bid_id}:
    delete:
      consumes:
      - application/json
      description: This endpoint allows a contractor to delete a specific bid they
        made.
      parameters:
      - description: Bid ID
        in: path
        name: bid_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted bid
          schema:
            type: string
        "400":
          description: Invalid bid ID or contractor ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a bid by contractor ID
      tags:
      - bids
  /api/contractor/tenders/{tender_id}/bid:
    post:
      consumes:
      - application/json
      description: This endpoint allows a contractor to create a new bid for a specified
        tender. The contractor must provide the bid details in the request body.
      parameters:
      - description: Tender ID
        in: path
        name: tender_id
        required: true
        type: string
      - description: Bid details
        in: body
        name: bid
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateBidRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created bid
          schema:
            $ref: '#/definitions/handlers.Bid'
        "400":
          description: Invalid tender ID or bad request body
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new bid
      tags:
      - bids
  /api/contractor/tenders/{tender_id}/bids:
    get:
      consumes:
      - application/json
      description: This endpoint retrieves a list of bids for a specified tender.
        The list can be filtered and sorted using query parameters.
      parameters:
      - description: Tender ID
        in: path
        name: tender_id
        required: true
        type: string
      - description: Filter by bid status
        in: query
        name: status
        type: string
      - description: Search bids by comments
        in: query
        name: search
        type: string
      - description: Minimum bid price
        in: query
        name: min_price
        type: number
      - description: Maximum bid price
        in: query
        name: max_price
        type: number
      - description: Minimum delivery time
        in: query
        name: min_delivery_time
        type: integer
      - description: Maximum delivery time
        in: query
        name: max_delivery_time
        type: integer
      - description: Sort by field
        in: query
        name: sort_by
        type: string
      - description: Sort order (asc or desc)
        in: query
        name: sort_order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of bids
          schema:
            items:
              $ref: '#/definitions/handlers.Bid'
            type: array
        "400":
          description: Invalid tender ID
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List bids for a tender
      tags:
      - bids
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"